pallet_membership::pallet

Struct Pallet

pub struct Pallet<T, I = ()>(/* private fields */);

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Implementations

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn add_member( origin: OriginFor<T>, who: <<T as Config>::Lookup as StaticLookup>::Source, ) -> DispatchResultWithPostInfo

Add a member who to the set.

May only be called from T::AddOrigin.

pub fn remove_member( origin: OriginFor<T>, who: <<T as Config>::Lookup as StaticLookup>::Source, ) -> DispatchResultWithPostInfo

Remove a member who from the set.

May only be called from T::RemoveOrigin.

pub fn swap_member( origin: OriginFor<T>, remove: <<T as Config>::Lookup as StaticLookup>::Source, add: <<T as Config>::Lookup as StaticLookup>::Source, ) -> DispatchResultWithPostInfo

Swap out one member remove for another add.

May only be called from T::SwapOrigin.

Prime membership is not passed from remove to add, if extant.

pub fn reset_members( origin: OriginFor<T>, members: Vec<T::AccountId>, ) -> DispatchResult

Change the membership to a new set, disregarding the existing membership. Be nice and pass members pre-sorted.

May only be called from T::ResetOrigin.

pub fn change_key( origin: OriginFor<T>, new: <<T as Config>::Lookup as StaticLookup>::Source, ) -> DispatchResultWithPostInfo

Swap out the sending member for some other key new.

May only be called from Signed origin of a current member.

Prime membership is passed from the origin account to new, if extant.

pub fn set_prime( origin: OriginFor<T>, who: <<T as Config>::Lookup as StaticLookup>::Source, ) -> DispatchResultWithPostInfo

Set the prime member. Must be a current member.

May only be called from T::PrimeOrigin.

pub fn clear_prime(origin: OriginFor<T>) -> DispatchResult

Remove the prime member if it exists.

May only be called from T::PrimeOrigin.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn members() -> BoundedVec<T::AccountId, T::MaxMembers>

The current membership, stored as an ordered Vec.

pub fn prime() -> Option<T::AccountId>

The current prime member, if one exists.