Trait pallet_membership::Trait

pub trait Trait<I = DefaultInstance>: Trait {
    type Event: From<Event<Self, I>> + Into<Self::Event>;
    type AddOrigin: EnsureOrigin<Self::Origin>;
    type RemoveOrigin: EnsureOrigin<Self::Origin>;
    type SwapOrigin: EnsureOrigin<Self::Origin>;
    type ResetOrigin: EnsureOrigin<Self::Origin>;
    type MembershipInitialized: InitializeMembers<Self::AccountId>;
    type MembershipChanged: ChangeMembers<Self::AccountId>;
}

Associated Types

type Event: From<Event<Self, I>> + Into<Self::Event>

The overarching event type.

type AddOrigin: EnsureOrigin<Self::Origin>

Required origin for adding a member (though can always be Root).

type RemoveOrigin: EnsureOrigin<Self::Origin>

Required origin for removing a member (though can always be Root).

type SwapOrigin: EnsureOrigin<Self::Origin>

Required origin for adding and removing a member in a single action.

type ResetOrigin: EnsureOrigin<Self::Origin>

Required origin for resetting membership.

type MembershipInitialized: InitializeMembers<Self::AccountId>

The receiver of the signal for when the membership has been initialized. This happens pre- genesis and will usually be the same as MembershipChanged. If you need to do something different on initialization, then you can change this accordingly.

type MembershipChanged: ChangeMembers<Self::AccountId>

The receiver of the signal for when the membership has changed.

Loading content...