Trait pallet_membership::pallet::Config

pub trait Config<I: 'static = ()>: Config {
    type Event: From<Event<Self, I>> + IsType<<Self as Config>::Event>;
    type AddOrigin: EnsureOrigin<Self::Origin>;
    type RemoveOrigin: EnsureOrigin<Self::Origin>;
    type SwapOrigin: EnsureOrigin<Self::Origin>;
    type ResetOrigin: EnsureOrigin<Self::Origin>;
    type PrimeOrigin: EnsureOrigin<Self::Origin>;
    type MembershipInitialized: InitializeMembers<Self::AccountId>;
    type MembershipChanged: ChangeMembers<Self::AccountId>;
    type MaxMembers: Get<u32>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self, I>> + IsType<<Self as Config>::Event>

The overarching event type.

type AddOrigin: EnsureOrigin<Self::Origin>

Required origin for adding a member (though can always be Root).

type RemoveOrigin: EnsureOrigin<Self::Origin>

Required origin for removing a member (though can always be Root).

type SwapOrigin: EnsureOrigin<Self::Origin>

Required origin for adding and removing a member in a single action.

type ResetOrigin: EnsureOrigin<Self::Origin>

Required origin for resetting membership.

type PrimeOrigin: EnsureOrigin<Self::Origin>

Required origin for setting or resetting the prime member.

type MembershipInitialized: InitializeMembers<Self::AccountId>

The receiver of the signal for when the membership has been initialized. This happens pre-genesis and will usually be the same as MembershipChanged. If you need to do something different on initialization, then you can change this accordingly.

type MembershipChanged: ChangeMembers<Self::AccountId>

The receiver of the signal for when the membership has changed.

type MaxMembers: Get<u32>

The maximum number of members that this membership can have.

This is used for benchmarking. Re-run the benchmarks if this changes.

This is enforced in the code; the membership size can not exceed this limit.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.